[PP] split journey controller into pieces
