+++ README.md
@@ -34,8 +34,7 @@ sbt> runTestOnly
 Journey states correspond to the result of the submission on pages (or endpoints).
 
 ```mermaid
-
-flowchart TD
+flowchart TD;
 
     govukpage[\Claim an income tax refund a gov.uk page\] 
     -- User clicked `/start` link on gov.uk page.
