+++ app/action/JourneyIdKey.scala
@@ -1,0 +1,5 @@
+package action
+
+object JourneyIdKey {
+  val journeyIdKey: String = "p800-refunds-frontend.journeyId"
+}
