+++ app/controllers/StartController.scala
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package controllers
+
+import action.JourneyIdKey
+import play.api.mvc._
+import services.JourneyService
+import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
+
+import javax.inject.{Inject, Singleton}
+import scala.concurrent.ExecutionContext
+
+@Singleton
+class StartController @Inject() (
+    mcc:            MessagesControllerComponents,
+    journeyService: JourneyService
+)(implicit ec: ExecutionContext) extends FrontendController(mcc) {
+
+  val get: Action[AnyContent] = Action.async { implicit request =>
+    journeyService
+      .newJourney()
+      .map(journey => Redirect(
+        routes.DoYouWantToSignInController.get
+      ).addingToSession(JourneyIdKey.journeyIdKey -> journey.id.value))
+  }
+
+}
