+++ app/views/CheckYourReferencePage.scala.html
@@ -27,7 +27,7 @@
 )
 
 @(
-    reference: String,
+    reference: P800Reference,
     form: Form[CheckYourReferenceFormValue],
     submitUrl: Call
 )(implicit request: Request[_])
@@ -42,13 +42,17 @@
     "reference-check"
 }
 
-@layout(pageTitle = Some(heading), backLinkUrlOverride = Some(controllers.routes.JourneyController.enterP800Reference.url)) {
+@layout(
+    pageTitle = Some(heading),
+    backLinkUrlOverride = Some(controllers.routes.EnterP800ReferenceController.get.url)
+) {
+
     @if(form.errors.nonEmpty) {
         @viewsHelpers.govukErrorSummary(ErrorSummary(errorList = form.errors.asTextErrorLinks, title = Text(Messages.CommonMessages.`There is a problem`.show)))
     }
 
     <h1 class="govuk-heading-xl">@heading</h1>
-    <p class="govuk-body">@{Html(Messages.CheckYourReferenceMessages.`You entered X`(reference).show)}</p>
+    <p class="govuk-body">@{Html(Messages.CheckYourReferenceMessages.`You entered X`(reference.value).show)}</p>
 
     <h2 class="govuk-heading-m">@{Messages.CheckYourReferenceMessages.`Is this correct?`.show}</h2>
