+++ app/views/EnterP800ReferencePage.scala.html
@@ -42,7 +42,11 @@
     Messages.EnterP800ReferenceMessages.`What is your P800 reference?`.show
 }
 
-@layout(pageTitle = Some(title), backLinkUrlOverride = Some(controllers.routes.JourneyController.doYouWantToSignIn.url)) {
+@layout(
+    pageTitle = Some(title),
+    backLinkUrlOverride = Some(controllers.routes.DoYouWantToSignInController.get.url)
+) {
+
     @if(form.errors.nonEmpty) {
         @viewsHelpers.govukErrorSummary(ErrorSummary(errorList = form.errors.asTextErrorLinks, title = Text(Messages.CommonMessages.`There is a problem`.show)))
     }
