+++ app/views/WeNeedYouToConfirmYourIdentityPage.scala.html
@@ -31,7 +31,10 @@
     Messages.ConfirmIdentity.`We need you to confirm your identity`.show
 }
 
-@layout(pageTitle = Some(title), backLinkUrlOverride = Some(controllers.routes.JourneyController.doYouWantToSignIn.url)) {
+@layout(
+    pageTitle = Some(title),
+    backLinkUrlOverride = Some(controllers.routes.DoYouWantToSignInController.get.url)
+) {
 
     <h1 class="govuk-heading-xl">@title</h1>
 
@@ -59,5 +62,4 @@
         content = Text(Messages.CommonMessages.`continue`.show),
         id      = Some("submit")
     ))
-
 }
