+++ app/views/chequejourney/RequestReceivedPage.scala.html
@@ -21,12 +21,15 @@
 @import views.html.helper.CSPNonce
 
 @this(
-        layout: Layout,
-        viewsHelpers: views.ViewsHelpers,
-        appConfig: AppConfig
+    layout: Layout,
+    viewsHelpers: views.ViewsHelpers,
+    appConfig: AppConfig
 )
 
-@(p800Reference: P800Reference, refundAmountInPence: AmountInPence)(implicit request: Request[_])
+@(
+    p800Reference: P800Reference,
+    refundAmountInPence: AmountInPence
+)(implicit request: Request[_])
 
 @import viewsHelpers.requestSupport._
