+++ app/views/testonly/GovUkStubPage.scala.html
@@ -34,6 +34,6 @@
   @viewsHelpers.govukButton(Button(
     content = Text("Claim Now >"),
     attributes = Map("id" -> "tryAgain"),
-    href = Some(controllers.routes.JourneyController.start.url)
+    href = Some(controllers.routes.StartController.get.url)
   ))
 }
